refactor(codex): extract goal orchestration into RuntimeGoalCoordinator
The 0.153 goal lifecycle work grew runtime.ts past the repository line
budget (2223 lines against the 1750 grandfathered ceiling). Moves /goal
command handling, queued goal-control delivery across turn transitions,
goal notification handling, and the goal set/replace/clear rollback
helpers into runtime-goal.ts behind a narrow RuntimeGoalPort exposing
the runtime-owned session-lifecycle operations (sessions map, finish,
pending-finish flush, turn-attempt admission). Pure move; behavior and
public API unchanged.
